Handle fread errors in sy_refill
Before this commit, the caller needed to call feof and ferror if
sy_refill did not return a value larger than the requested size. With
this commit, as long as the caller checks the error out parameter, they
can trust that the return value will only fail to meet the requested
size when they have reached the end of the stream.