Remove the state parameter from sy_token
I originally thought parsers would need the state variable to properly
parse tokens like string literals. While this is true if a parser aims
to parse everything with the same technique, it makes more sense to hand
off tokens like string literals to a separate function that uses a
different technique to parse.